Rename `exchangeRegistry` to `exchange-registry` for consistency
